NimbusU-backend/pyproject.toml
[project]
name = "nimbusu-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "python-decouple>=3.8",
    "boto3>=1.36.0",
    "django-storages>=1.14.0",
    "channels[daphne]>=4.2.0",
    "daphne>=4.1.2",
]
